data_processor/requirements.txt
kafka-python
rdflib